fix(suite): Update ready on section
# Conflicts:
#	packages/suite/src/components/suite/modals/ReduxModal/UserContextModal/ActivateAssetsModal.tsx
