+++ packages/product-components/src/components/TokenIconSet/TokenIconSet.tsx
@@ -1,16 +1,11 @@
 import { useMemo } from 'react';
 
-import styled, { css } from 'styled-components';
-
 import { type NetworkSymbol } from '@suite-common/wallet-config';
-import { Text } from '@trezor/components';
-import { type SpacingValuesNew, borders } from '@trezor/theme';
+import { type SpacingValuesNew } from '@trezor/theme';
 
-import { mapSizeToTypographyStyle } from './utils';
 import { AssetLogo } from '../AssetLogo/AssetLogo';
 import { type AssetLogoSize } from '../AssetLogo/AssetLogoWithId';
-
-const MAX_VISIBLE_TOKENS = 3;
+import { IconSetBase, IconWrapper, MAX_VISIBLE_ICONS } from '../IconSet/IconSetBase';
 
 export type TokenIconSetProps = {
     symbol: NetworkSymbol;
@@ -25,71 +20,6 @@ export type TokenIconSetProps = {
     reverseVisibleTokens?: boolean;
 };
 
-const Container = styled.div<{
-    $length: number;
-    $size: AssetLogoSize;
-    $gap: SpacingValuesNew;
-    $isCountVisible: boolean;
-    $isCentered: boolean;
-}>`
-    justify-content: center;
-    display: flex;
-    align-items: center;
-
-    ${({ $isCentered, $size, $gap, $length, $isCountVisible }) => {
-        const visibleCount = $length > 3 ? 3 + Number($isCountVisible) : $length;
-
-        return $isCentered
-            ? css`
-                  width: ${$size}px;
-              `
-            : css`
-                  width: ${$size + (visibleCount - 1) * $gap}px;
-              `;
-    }}
-
-    ${({ $length, $gap, $isCountVisible }) =>
-        $length > 1 &&
-        css`
-            display: grid;
-            grid-template-columns: repeat(
-                ${$length > 3 ? 3 + Number($isCountVisible) : $length},
-                ${$gap}px
-            );
-            justify-items: center;
-        `}
-`;
-
-const IconWrapper = styled.div<{ $size: number; $gap: number; $length: number }>`
-    border-radius: ${borders.radii.full};
-
-    ${({ $size, $gap, $length }) =>
-        $length > 1 &&
-        css`
-            &:not(:last-child) {
-                mask: radial-gradient(
-                    circle at calc(50% + ${$gap}px) 50%,
-                    transparent ${$size / 2 + 1}px,
-                    black ${$size / 2 + 1}px
-                );
-            }
-        `}
-`;
-
-const CountContainer = styled.div<{ $size: AssetLogoSize }>`
-    ${({ $size }) => css`
-        width: ${$size}px;
-        height: ${$size}px;
-    `}
-
-    position: relative;
-    display: flex;
-    align-items: center;
-    justify-content: center;
-    border-radius: ${borders.radii.full};
-    background: ${({ theme }) => theme.legacyBackgroundTertiaryDefaultOnElevationNegative};
-`;
-
 export const TokenIconSet = ({
     symbol,
     tokens,
@@ -102,7 +32,7 @@ export const TokenIconSet = ({
     const { length } = tokens;
 
     const visibleTokensContent = useMemo(() => {
-        const visibleTokens = tokens.slice(0, MAX_VISIBLE_TOKENS);
+        const visibleTokens = tokens.slice(0, MAX_VISIBLE_ICONS);
         const orderedTokens = reverseVisibleTokens ? visibleTokens.reverse() : visibleTokens;
 
         return orderedTokens?.map(token => (
@@ -118,26 +48,15 @@ export const TokenIconSet = ({
         ));
     }, [tokens, reverseVisibleTokens, symbol, size, gap, length]);
 
-    if (length === 0) {
-        return null;
-    }
-
     return (
-        <Container
-            $length={length}
-            $size={size}
-            $gap={gap}
-            $isCountVisible={isCountVisible}
-            $isCentered={isCentered}
+        <IconSetBase
+            count={length}
+            size={size}
+            gap={gap}
+            isCountVisible={isCountVisible}
+            isCentered={isCentered}
         >
             {visibleTokensContent}
-            {length > 3 && isCountVisible && (
-                <CountContainer $size={size}>
-                    <Text typographyStyle={mapSizeToTypographyStyle(size)} intent="neutral">
-                        +{length - MAX_VISIBLE_TOKENS}
-                    </Text>
-                </CountContainer>
-            )}
-        </Container>
+        </IconSetBase>
     );
 };
