+++ packages/suite/src/components/suite/modals/ReduxModal/UserContextModal/ActivateAssetsModal.tsx
@@ -135,7 +135,7 @@ export const ActivateAssetsModal = ({ onCancel }: ActivateAssetsModalProps) => {
                             isDisabled={isApplyPending}
                             data-testid="@modal/activate-assets/save"
                         >
-                            <Translation id="TR_SAVE" />
+                            <Translation id="TR_CONFIRM" />
                         </Modal.Button>
                     ) : null
                 }
